v-assets: refactored seed templates into new location
